Adding Databricks Notebooks
Adding the bronze-to-silver , silver-to-gold, and storage-mount notebooks.

storage_mount notebook contains the code to mount bronze, silver, gold directories in Azure Data Lake Storage Gen2. This part is already done. No need to run this notebook again.
 
The bronze-to-silver , silver-to-gold notebooks both contain code to transform the datasets into silver level and gold level respectively.
These are essential part of the ETL pipeline